Addressing code review findings
